Add checks for idle dupes inside of busy rockets
This change adds a check for idle duplicants inside of travelling or
mining rockets and will squelch Idle notices from these duplicants by
default. Additionally the change adds a module option to control this
behavior. Finally, when the PauseOnIdle module option is set the
change will also defer pausing until after the notification is
displayed to the screen.