Add a configurable pause cooldown
When the PauseOnIdle option is enabled, this commit adds a separate
configurable cooldown time to prevent multiple sequential idle
notifications from immediately pause the game repeatedly. This enables
the player to unpause for _cooldown_ seconds before triggering another
pause if a new idle notification is generated.

Additionally bump minimumSupportedBuild to 545357 as this was the most
recent version the mod was tested against.