Initial cut of the CGSM mod
CGSM, or Cluster Generation Settings Manager, is an ONI mod which
allows the player to control how the starmap is initially generated
when creating a new game. This includes which planetoids to include or
exclude from the starmap, the size of the starmap, which space POIs to
include or exclude, the number of artifact POIs, and more.

Only the mini-cluster planetoids, marshy planetoid, fire planetoid,
water planetoid, regolith planetoid, tundra planetoid, and moo
planetoid are supported in this first release. Subsequent releases
will add more planetoids, allow controllling their size, and
potentially their geyser counts.