CGSM: Decouple core logic from Option input & YAML output
This commit is a pure refactor decoupling Option input parsing and
YAML output emission from the internal representation of a
cluster. This enables the basic logic to be used outside of these
contexts and sets the stage for adding additional features such as
adding randomization or constraints. (e.g. ensuring that player has
pathways to atmo suits and isn't in a dead-end).