CGSM: Refactor for existing game clusters (part 1)
This commit is a refactor only in prep for supporting using the game's
existing clusters as a starting template for
customization. Specifically this change completes the decoupling of
Options from Cluster and introduces new types PlanetoidPlacement,
ArtifactPOIs, POIType, and POIGroup. Cluster and ClusterYamlEmitter
have been refactored to use these new types.