CGSM: Add support for existing game clusters
This commit adds support for generating a customized cluster using one
of the game's built-in clusters as a starting template. Built-in
clusters can be customized via the game settings button when launching
a new colony (same place as WGSM). Initially only adding or removing
planetoids from built-in clusters is supported. Customizing the
starmap radius and POIs for built-in clusters will be added in a
future release.