CGSM: Import of phan's PlibOptions@380c52938 (part 1)
This commit copies verbatim ONIMods/PLibOptions locally because we
want to make a few minor changes to it. This includes adding a
RequireMod attribute and extending this & the existing RequireDLC
attribute to apply to Enums. (Currently RequireDLC only applies to
non-Enum option entries). This commit is an import only; our local
changes will follow in subsequent  commits.