CGSM: Introduce RequireMod local PlibOptionsFilt attribute
This commit introduces a RequireMod attribute for our local
PlibOptionsFilt. The RequireMod() attribute may be added to any
Plib.Option field or enum value in order to conditionally include the
optiona/enum value. When the specified mod is installed & enabled,
these options/enum values will be available for selection in the mod
options menu and otherwise  not.