CGSM: Add support for Roller Snakes Reborn
This commit adds support for the 2 clusters and 6 planetoids
introduced by the Roller Snakes Reborn mod. Tetrament and
TetramentVanilla can now be selected as start planetoids. Dry
Radioactive Forest can now be selected as a warp planetoid, and
the Snakes Reborn customized versions of the Marshy, Water, and Fire
planetoids can be selected as optional  planetoids.