CGSM: Add support for MiniBase mod
This commit adds support for including MiniBase Spaced Out planetoids
in a CGSM cluster or when customizing an existing cluster. Support for
customizing the MiniBase Spaced Out cluster itself will be added in a
subsequent release. Note that as of the time of this commit these
changes will not work with the currently released version of the
aMiniBase mod due to an incompatibility in the way MiniBase Spaced Out
customizes world generation. I have submitted a PR to the author of
the MiniBase Spaced Out mod to fix this
(https://github.com/antoine-fresse/MiniBaseSO/pull/2).