CGSM: add support for Skewed asteroid & cluster
This commit adds support for selecting the Skewed asteroid as a
starting planetoid in a CGSM custom cluster. Additionally, this commit
adds support for customizing the game's new built-in Klei Fest 2023
cluster.

While testing I also discovered that one of the updates broke our
SetQualitySetting() hook in some circumstances leading to finnicky
behavior when toggling planetoids in the game settings. This is fixed
as well in this commit.