BeyondExtraHard: add new mass toggle
This commit adds a new Beyond Extra Hard game setting to control the
amount of mass returned by digging or deconstructing. Normally ONI
will return 50% of the mass of a natural tile when dug and 100% of the
masses of materials in a deconstructed building. The "Restocking Fee"
setting will reduce these to 40% and 80% respectively and the
"Confiscatory Taxation" setting will reduce these to 30% and 60%
respectively.