QuickStart: fix crash when multiple new games are created in same session
This commit fixes a crash reported by user MilesObrienWasAUnionMan. We
failed to reset GameState.needDeferredDupeUpgrade when a new game is
loaded. This won't impact loading existing saved games, but if 2 new
games are created in the same session of oni this will trigger the
crash. This commit also adds a safety check for a missing attribute in
DupeBufs.upgradeAttributeLevels().