QuickStart: fix printing pod crash on new games
When a player creates a new game, plays it for some period of time,
and then attempts to print a new duplicant from the printing pod this
triggers a crash. This crash is not triggered if the game is instead
loaded from a .sav file. The root cause is that
MinionIdentity.OnSpawn() is invoked for both dupes present in a colony
and potential dupes that may be chosen from the
ImmigrantScreen. Previously we were not differentiating between them
and were attempting to upgrade the attributes of a dupe displayed on
the ImmigrantScreen, which crashes in AttributeLevel.LevelUp() because
of uninitialized dependencies. This commit addresses this.

Additionally this commit prevents upgrading new dupes; we only upgrade
the initial set of dupes.